Add systemd socket activation support.
Since systemd was introduced lirc clients have had problems at
startup when trying to connect to the lircd socket before it's
created. The root cause is the aggressive parallell boot performed
by systemd.

The solution is to add socket activation. In this mode, the socket
is created by systemd and handed over to lircd at startup. The patch
implements this. It should be reasonably transparent.

At configure time it enables systemd support if it's available,
otherwise it's silently ignored.

In runtime lircd looks for and uses a socket handed over by systemd.
If there is no such socket it proceeds as normal.

Adds a dependency on pkg-config for the PKG_CHECK_MODULES macro.

Reviewed at systemd devel list:
http://lists.freedesktop.org/archives/systemd-devel/2013-October/013648.html