daemons, tools: Build against new library.
There are at least four more or less copy-pasted waitfordata()
functions out there. The one in lircd i more elaborated, though.
Picking the one in irrecord, which looks the same as the ones
in mode2 and xmode2. Placing function in receive, which looks
natural. Moving stuff from daemons and tools to library.