doc: Ditch help2man, use traditional manpage sources.
The current manpage generation is a mess. The dependencies are unclear,
and the generated files are not flexible enough to create descent
manpages. Also, creating manpages for non-commands e. g., config files
requires wrappers. All in all, a mess.

Remove release-html and release-man scripts, do everything in the
Makefile.am. Check in current generated manpages as source and fix them
up.

This also opens up for traditional keyword substitution in the generated
files. Currently only version is patched, but this is easily expanded.

Removing generation of table.html and lirc.hwdb for now, to be
re-enabled later when there is new configuration system in place.