DEBUG: Remove unneeded configuration option.
All debug output is now governed by run-time options. Having #ifdef DEBUG
creates many problems when users needs to recompile lirc whne running
into bugs. There is a performance hit, but it's basically just an integer
comparison when output is disabled.

Builds not defining DEBUG are also likely to be untested sine most devs
always have it on. On the contrary, many distros have it disabled, so
mosgt users have it disabled. This is not a reasonable situation.