Remove unnecessary and silly sorting of remotes (#28).
The idea is probably to make the program faster in that remotes with
"shorter" codes comes at the beginning, and thus (?) leads to faster
decoding. However, this idea is both badly implemented (raw remotes
comes out with a bit length of 0, so they will always be at the
beginning), and is also counter productive, because the only sensible
performance related optimization of lircd.conf is when the user manually
reorders the file with more commonly used remotes first. The (for the
user invisible!) re-ordering undoes his careful work!

Fix in code consists of just removing the sorting.