plugins: Add irtoy driver (#20).
Adds the plugins/irtoy.c driver for the Dangerous
Prototypes USB Infrared Toy 2.  Original patch information created
by Peter Kooiman.  I have read references that it might not be as
up-to-date as desired due to "robustness" changes in the transmit
command protocol.  This current set works to transmit and receive.
It even works with the irrecord utility.

This has been tested with the "latest" v22 firmware images on the
IRToy2 as well as with a "beta" (?) v23 one that no one seems to
be supporting, other than to make it available.