logging: Enable symbolic constants in UI, clean up.
Users should be able to set the loglevel using the familiar
'info', 'warning' etc. Add interfaces in lirc_log to parse
loglevels and in options to store them. New code in client's
option parsing + various cleanup of leftovers from the time
debug was a boolean, not a level. Adding some tests.