lib: lirc_client: Add ignore_first_events i. e., longpress option (#50).
This option controls the amount of repeats before delay period activates.
Default value is 0, which makes it what it was before: delay period was
starting after 1 event. If set to 1 (true) the first event is suppressed
which makes it is possible to define the reaction to long key-presses.

Ignore "ignore_first_event" parameter if "delay" is not set
and warn the user about that misconfiguration.

Also make rep_filter() much more readable.