lib: Add new lircrc_class global lircrc option (#61).
Adds a new global option lircrc_class to lircrc format. The
format is simple

    lircrc_class <string>

The string is used as argument when creating lircrc sockets. The
shebang is parsed as a lircrc_class <basename filename> i. e.,
it emulates the old behaviour to create a socket named after
the lircrc config file.

The shebang is deprecated and usage gives  a warning.

lirc_class can only be used in the toplevel config file, not
in included ones. If found in an included file it generates
a warning but is otherwise ignored. Same goes for bad syntax.

Fixing configure.html + some tests.