lib, lircrcd: Move lircrcd sockets to /var/run (#62).
This is not completely compatible with the old behaviour: previously, if
the files had the same paths but different basenames they got different
lircrd sockets. This patch makes all lircrc files with the same basename
use the same socket. Can't think of any usecase where this is a problem

As of now we also support lircrc_class which can be used to walk around
this limitation.

The old code actually allowed the use of other commands than lircrcd as
interpreter in the shebang. This was never documented or an official
interface. It's not supported with this patch.

Patch solves all sorts of problems related to using sockets in the same
directory as config files.