srm7500: Fix remote not working on newer versions of libusb.
A couple of months ago I noticed, after a libusb update, that it
was no longer working, constantly outputting to /var/log/messages the
'could not detach kernel driver' error.

Upon further digging between lirc and libusb, I noticed that error codes
changed recently, and lirc was entering the if that triggers that error
due to libusb returning another type of error (EINVAL) instead of the
original one (ENODATA). The patch treats both of these the same way.