configs: Nove obsolete configs to attic
The lirc_gpio driver is retired since long. Moving configs
depneding on this driver to new "attic" place so they can
still be found if someone wants to resurrect them.

The wpc8769l driver didn't make it when the lirc drivers were
upstreamed to the kernel. Moving the config to the attic, and
re-adding the kernel driver (which doesn't build as-is) if
someone want's to resurrect them.