build: Add selinux policy.
The selinux policy is built if the libselinux-devel files (really
libselinux.pc) is present. The set of policies (mls, targeted) is
hardcoded in Makefile.am.  Although polices are built they are not
installed during 'make install'.

The long time solution is to add this policy to the upstream generic
one. However, using a package policy yields a faster turn-around in
the initial incarnations. There is also some problems with the
module naming:

https://github.com/fedora-selinux/selinux-policy/issues/38