build:  Debian packaging build wrap-up.
Locating modinfo is causing some pain on Debian system which havn't
/sbin:/usr/sbin in $PATH. Locate using autoconf and use that value.
While we are on it, rename paths.py to config.py, which reflects the
purpose.

A VERSION file is added, installed as a document. Git version data
is updated at "make dist".

The debian packaging is updated, and outside the package the Fedora
debian toolchain has been updated so that source packages can be built.

Various bugfixes and clean-up.