Clean up generating of devinput.lircd.conf (#153).
The existing logic was completely broken. The idea to distribute
input_map.inc is wrong since it reflects the kernel on the machine
where it was created. Better then to just distribute a best-effort
devinput.lircd.conf + simple tool to re-generate it from the target
system kernel.

After patch there is just one single script used both when building and
when users re-generate devinput.lircd.conf.