Add plugin device enumeration.
The only reliable way to enumerate devices, in particular the USB ones
is done by the driver. This adds a new option in the configs file +
a mode2 CLI option to display the values listed by the driver. The
actual code is a DRVCTL so it does not really change the driver
interface.

Implement in atilibusb and ftdi, the latter untested. The ftdi code
fixes an existing error in the drvctl_func.