plugins/ftdi: Tear down subprocess if parent dies
If the lircd parent process is killed with SIGKILL it will be unable
to call `hwftdi_deinit` and so the subprocess it spawned will continue
living, keeping the ftdi device busy.  This change uses receiving EOF
on the pipe from the parent process as a signal that the parent process
has died and will shut itself down accordingly too.