ftdi: Add ftdix driver
Mode2 transmitter using the bitbang mode of an FTDI USB-to-serial chip such as
the FT230X.  The standard FTDI driver with the FT232R is unreliable having a
IR transmit failure rate of ~2%.  The FT232R also becomes unusable if you set
the tx_baud_rate to values other than 65536.

This driver works differently modifying its tx_baud_rate according to the
carrier frequency set.  This is ok because the FT230X does not have the timing
bugs present in the FT232R.  This allows much lower data rates over USB and
so improves reliablity, particularly if other devices are connect to the same
USB port.

Receive is not included to significantly simplify design of this driver.

Having this driver in the same file as the ftdi driver makes sharing code
easier and avoids confusing plugins/make-pluginlist.sh which is called during
autogen.