plugins/ftdi: Specify duty_cycle to fix ftdi driver
The recent commit c9afda9 added support for duty_cycle to the ftdi driver.
This worked for lircd config files that actually specified `duty_cycle`
but broke it for config files that didn't as it would default to 0 rather
than 50.

Before this commit it was the responsibility of each driver that implements
duty_cycle to know about this default.  This fix centralises the defaults
by adding a getter function to `ir_remote`, simplifying the drivers.  As a
result, even though this patch is to fix a bug in ftdi it touches the
default driver too.

This issue didn't come up in testing c9afda9 because I'd only tested it
with config files that actually explicitly set duty_cycle.

For the ftdix driver I hard-code a duty_cycle of 50 because that's all it's
capable of.  Passing in duty_cycle to modulate_pulses in this instance was
misleading.