plugins/uirt: Don't block if a 0xff is received by usb-uirt .
Periodically the usb§$-uirt-raw sends a short signal without any IR
activity. I don't know what the signal is, but as it's a "raw" receiver
I presume it's some interference.

Regardless, the problem is that the uirt receiver code sees this short
signal, and the main loop goes back to the uirt2 code because there is
data remaining to be read. It reads a 0xff as the first byte in
uirt2_read_raw and that function goes back to read more .. but there
isn't any more, so it blocks forever because the main loop function gave
us a timeout of zero .. which is turned into an infinite timeout in
waitfordata.

Subsequent IR keypresses DO work, because the blocking read gets some
data. But subsequent connections to the socket (e.g. by irw or irsend)
DON'T work: they block until some IR activity occurs which could be many
minutes later, because lircd is stuck in a poll on the USB fd, but isn't
polling on the socket connection fd, which is only done by the main
loop.

There may other ways to fix this, but this does work. With the patch I
can always connect to the socket, and IR keypresses continue to be
processed correctly.