ftdi: Close uninteresting FDs in subprocess
The ftdi driver forks a subprocess to interact with the ftdi device itself.
This means that the file-descriptor table is duplicated into the subprocess.
This includes the file-descriptors corresponding to the socket client
connections meaning if the parent process dies the clients wont be notified as
long as the child process is still around.

I think I've been seeing some crashes in lircd, but the clients are just timing
out rather than erroring.  This should make things clearer.

This is clearly not perfect as if there are more than 20 FDs open in lircd some
files may still be open in the subprocess, but it does make things a little
cleaner for the purposes of debugging.