ftdi: Fix retry logic
The ftdi driver retries if setting up the hardware fails.  This logic had
a bug where if something were to go wrong after opening the device the
driver would then attempt to re-open the device while it was still open.
This lead to the reopening failing and another retry: an infinite loop.

This commit fixes that issue by performing open retry logic seperately
from all the post-open retry logic.