lib: use proper linking method to avoid parallel build issue
Using <foo>_LDFLAGS = -l<library> is correct when <library> is an
external library. However, when it is built by the same package, and
especially in the same directory, this is wrong and can cause parallel
build issues. In lib/Makefile.am, there was:

libirrecord_la_LDFLAGS      = -llirc

But the liblirc library is built in the same directory. Or, due to the
using of <foo>_LDFLAGS, make is not aware of the build dependency
between libirrecord and liblirc.

To solve this, <foo>_LIBADD should be used instead, as follows:

libirrecord_la_LIBADD       = liblirc.la

This fixes parallel build issues seen by automated build tests
conducted by the Buildroot project, such as:

  http://autobuild.buildroot.org/results/eb4/eb47d57de8182d25b1dacbf0ac3726ed20063d04/build-end.log

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>