ftdix: Support opening FTDI device by path to device node
This will be helpful to distinguish between FTDI devices when there are
multiple of them attached to the same machine.  You can use udev to create
a well known symlink name pointing at the `/dev/bus/usb/BBB/DDD` device
node and pass that as `--device` to lirc.

This isn't perfect: If you're not using udev and you have created a device
node yourself (with mknod) whose name doesn't end in `/BBB/DDD` then this
will fail.  Fixing that edge case would require patching libusb, but it
shouldn't come up in practice.

We still support the alternative syntax of identifying a device by
product/vendor/serial.  We can distinguish between the two syntaxes as
absolute filenames always start with a `/`.