Build: Add --enable-uinput ./configure option  (#238).
There are situations where the default value guessed from the
existence of the /dev/uinput device cannot be used to determine
if lircd-uinput should be built, all related to cross-compiling.
The bug occurs on Ubuntu, where some build hosts seems to lack
this device. It will also bite anyone doing a traditional
cross-compilation. The new option should work in all cases, not
adding so much  complexity since the default is almost always
sane.