python-bindings: feature complete.
Old and new python code now lives in  python-pkg/. This is installed
in  regular locations by Makefile, but it is also a functional
pypi package which could be installed e. g., in a virtualenv.

The API contains pieces to send commands and receive data. See the
API docs and also the manual. It is supposed to be complete in the
sense that it can handle all known socket commands to lircd(0) and
lircrcd(8). It can also receive both raw events like irw(1) and
application string like ircat(1).

There are also some proof-of-concepts in doc/python-bindings. If
installed in a virtualenc the pypi package installs the lirctool
script, a more complete demonstration of the API.