lircrcd: Rewrite to match manpage functionality (#253).
The old lircrcd implementation is plain broken with respect to the
manual page. It opens a connection ti lircd, listens to all decoded
keypresses and uses this to generate application strings to
clients.

New implementation is just s common instance and state for converting
codes to application strings. It only listens to clients, and
translates as required.

Caching key/mode transltions could make this daemon considerably faster.