Add lirc-cfg2csv tool
I was manually converting a lirc raw_codes remote control to non-raw to
hopefully improve reliability.  I wanted to check that I'd done it
correctly and investigate what effect my further changes would have on the
actual IR output.

I wanted to visualise this as I didn't have access to the actual IR
hardware (sender or receiver) so I was not in a position to fiddle and
iterate.  This way I can stick the CSV into a spreadsheet and plot the
codes themselves.

The `modulate_pulses` function is duplicated from the ftdi/ftdix driver.