lib: Fix crash when loading remote without "codes" block
When loading a config file that has a remote without "codes" block
(at least) irrecord crashes.

It turns out that sanityChecks() function tries to deference rem->codes
pointer blindly.
Fix it by checking whether this pointer is null and failing the validation
in this case.

Let's use this opportunity to also fix a logging call few lines earlier in
this function which tried to output a struct type as a string.