lib: Fix spinning in array_guest_code_equals()
lircd was spinning on startup in array_guest_code_equals() when started
with one of my configs.
That function didn't correctly advance current element pointers when
iterating over code lists in keys that had more than one code.

Fix this and also improper return value when the first argument key codes
are a prefix of the second argument ones (previously the function would
return "equal" verdict in this case).