build: install media/lirc.h (BTS #872074).
The plugin builds references this file. Basically, this should be
resolved by lirc_driver.h, testing if the kernel header is
available and using it if so, but without including config.h

This fix adds an evil symlink to not remove anything. The proper
fix includes actually moving the file and also streamlining the
HAVE_KERNEL_LIRC_H to the generic HAVE_LINUX_LIRC_H.