Only enable the lirc protocol if present
linux kernel 4.16 and newer no longer have the lirc chardev connected as
a protocol decoder, but it is baked into rc-core. So, it is not necessary
to enable the lirc protocol any more if lirc is not present in the
protocols file.

Note that lirc is still in the protocols file in current versions of the
kernel (at the time of writing), but it would be nice to be able to remove
this in the distant future.