mywaitfordata: Fix race condition
LIRC code uses SIGALRM for timers via `setitimer`, etc.  The signal
handlers set flags to say that they have occurred.  There was a race
condition between checking the flag and calling `poll` to block that could
cause lirc to hang until another event woke it up.

We use `ppoll` to avoid this race.  The signals are blocked before we check
the flags and atomically unblocked during the `ppoll` - so if a signal
arrives between the two the `ppoll` is gauranteed to be woken.