XMP: Fix transmitting with pre_data and post_data
`send_data` is called by `send_pre` and `send_post`, so it should always
end on a space rather than a pulse, otherwise you get incorrect
double-width pulses.  Instead the final pulse is intended to be written by
`send_trail`.

This fixes the implementation such that this is the case.  For backwards
compatibility with XMP configs that lack ptrail this also sets ptrail for
all XMP remotes.